refactor(accounts): give each accounts concern its own file
handler/admin/accounts.go had grown to 1726 lines holding route
registration, the handler struct, two caches, and every create / login /
verify / rebind / update / delete / batch / list / snapshot / balance
handler in one file, so any change to one concern showed up as a diff
against all of them.

Split by concern, same package, moves only:

- accounts.go          route registration, handler struct, caches (110)
- accounts_auth.go     login, verify token, rebind session, credential
                       normalization and recovery helpers (447)
- accounts_create.go   create + single-account create outcome (427)
- accounts_list.go     list, pagination, filters, snapshot computation (258)
- accounts_write.go    update, delete, batch (479)
- accounts_health_refresh.go  health refresh result types, balance (64)

The 28 declared funcs are byte-identical before and after; no signature,
receiver, or call site changed and no abstraction was added.
